// frontend/src/app/core/enums/tax.enum.ts
import { AccountTransferenceTax } from "../viewmodel/accountTransferenceTax.viewmodel";

// maxDaysPerTax
export enum TaxMaxDaysEnum{
	A = 1,
	B = 10,
	C = 20,
	C2 = 30,
	C3 = 40,
	C4 = 0
}
// maxValuePeerTax
export enum TaxMaxValueEnum{
	A = 1000,
	B = 2000,
	C = 0,
	C2 = 0,
	C3 = 0,
	C4 = 0
}
// taxPercent
export enum TaxPercentEnum{
	A = 3,
	B = 0,
	C = 8.2,
	C2 = 6.9,
	C3 = 4.7,
	C4 = 1.7
}
// taxValue
export enum TaxValueEnum{
	A = 3,
	B = 12,
	C = 0,
	C2 = 0,
	C3 = 0,
	C4 = 0
}